[tabs-stories-structure] move scenarios to snapshot testing dir
