+++ __docs__/wonder-blocks-tabs/navigation-tab-item.stories.tsx
@@ -9,13 +9,6 @@ import {Popover, PopoverContent} from "@khanacademy/wonder-blocks-popover";
 import Tooltip from "@khanacademy/wonder-blocks-tooltip";
 import {semanticColor, sizing} from "@khanacademy/wonder-blocks-tokens";
 import argTypes from "./navigation-tab-item.argtypes";
-import {
-    longText,
-    longTextWithNoWordBreak,
-} from "../components/text-for-testing";
-import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
-import {IconMappings} from "../wonder-blocks-icon/phosphor-icon.argtypes";
-import {ScenariosLayout} from "../components/scenarios-layout";
 
 const StyledUl = addStyle("ul", {
     margin: sizing.size_0,
@@ -137,87 +130,5 @@ export const CustomStyle: StoryComponentType = {
         style: {
             backgroundColor: semanticColor.surface.secondary,
         },
-    },
-};
-
-/**
- * The following story shows how the component handles specific scenarios.
- */
-export const Scenarios: StoryComponentType = {
-    render() {
-        const scenarios = [
-            {
-                name: "Long Text",
-                props: {
-                    current: true,
-                    children: <Link href="#link">{longText}</Link>,
-                },
-            },
-            {
-                name: "Long Text with No Word Break",
-                props: {
-                    current: true,
-                    children: (
-                        <Link href="#link">{longTextWithNoWordBreak}</Link>
-                    ),
-                },
-            },
-            {
-                name: "Long Text (icons)",
-                props: {
-                    current: true,
-                    children: (
-                        <Link
-                            href="#link"
-                            startIcon={
-                                <PhosphorIcon
-                                    icon={IconMappings.cookie}
-                                    size="small"
-                                />
-                            }
-                            endIcon={
-                                <PhosphorIcon
-                                    icon={IconMappings.iceCream}
-                                    size="small"
-                                />
-                            }
-                        >
-                            {longText}
-                        </Link>
-                    ),
-                },
-            },
-            {
-                name: "Long Text with No Word Break (icons)",
-                props: {
-                    current: true,
-                    children: (
-                        <Link
-                            href="#link"
-                            startIcon={
-                                <PhosphorIcon
-                                    icon={IconMappings.cookie}
-                                    size="small"
-                                />
-                            }
-                            endIcon={
-                                <PhosphorIcon
-                                    icon={IconMappings.iceCream}
-                                    size="small"
-                                />
-                            }
-                        >
-                            {longTextWithNoWordBreak}
-                        </Link>
-                    ),
-                },
-            },
-        ];
-
-        return (
-            <ScenariosLayout scenarios={scenarios}>
-                {(props) => <NavigationTabItem {...props} />}
-            </ScenariosLayout>
-        );
     },
 };
